+++ src/Machina/Robot.cs
@@ -102,7 +102,7 @@ namespace Machina
         /// <returns></returns>
         static public Robot Create(string name, RobotType make)
         {
-            if (!Util.IsValidVariableName(name))
+            if (!Utilities.Strings.IsValidVariableName(name))
             {
                 Machina.Logger.Error($"\"{name}\" is not a valid robot name, please start with a letter.");
                 return null;
@@ -1056,7 +1056,7 @@ namespace Machina
         /// <returns></returns>
         public bool DefineTool(Tool tool)
         {
-            if (!Util.IsValidVariableName(tool.name))
+            if (!Utilities.Strings.IsValidVariableName(tool.name))
             {
                 logger.Error($"\"{tool.name}\" is not a valid tool name, please start with a letter.");
                 return false;
@@ -1075,7 +1075,7 @@ namespace Machina
         /// <returns></returns>
         public bool DefineTool(string name, Point TCPPosition, Orientation TCPOrientation)
         {
-            if (!Util.IsValidVariableName(name))
+            if (!Utilities.Strings.IsValidVariableName(name))
             {
                 logger.Error($"\"{name}\" is not a valid tool name, please start with a letter.");
                 return false;
@@ -1096,7 +1096,7 @@ namespace Machina
         /// <returns></returns>
         public bool DefineTool(string name, Point TCPPosition, Orientation TCPOrientation, double weightKg, Point centerOfGravity)
         {
-            if (!Util.IsValidVariableName(name))
+            if (!Utilities.Strings.IsValidVariableName(name))
             {
                 logger.Error($"\"{name}\" is not a valid tool name, please start with a letter.");
                 return false;
@@ -1128,7 +1128,7 @@ namespace Machina
             double tcp_vX0, double tcp_vX1, double tcp_vX2, double tcp_vY0, double tcp_vY1, double tcp_vY2,
             double weight, double cogX, double cogY, double cogZ)
         {
-            if (!Util.IsValidVariableName(name))
+            if (!Utilities.Strings.IsValidVariableName(name))
             {
                 logger.Error($"\"{name}\" is not a valid tool name, please start with a letter.");
                 return false;
@@ -1152,7 +1152,7 @@ namespace Machina
         /// <returns></returns>
         public bool AttachTool(string toolName)
         {
-            if (!Util.IsValidVariableName(toolName))
+            if (!Utilities.Strings.IsValidVariableName(toolName))
             {
                 logger.Error($"\"{toolName}\" is not a valid tool name, please start with a letter.");
                 return false;